+++ Content.Server/_Exodus/Chemistry/CyclingHyposprayComponent.cs
@@ -1,10 +1,10 @@
 using Content.Shared.Chemistry.Reagent;
 using Robust.Shared.Prototypes;
 
-namespace Content.Server._Exodus.Silicons.Borgs;
+namespace Content.Server._Exodus.Chemistry;
 
 [RegisterComponent]
-public sealed partial class BorgParamedicHypoComponent : Component
+public sealed partial class CyclingHyposprayComponent : Component
 {
     [DataField]
     public string SolutionName = "hypospray";
