+++ Content.Server/_Exodus/Chemistry/CyclingHypospraySystem.cs
@@ -1,4 +1,3 @@
-using Content.Server._Exodus.Chemistry;
 using Content.Server.Chemistry.EntitySystems;
 using Content.Shared.Chemistry.Components;
 using Content.Shared.Chemistry.EntitySystems;
@@ -11,9 +10,9 @@ using Content.Shared.Mobs.Components;
 using Content.Shared.Popups;
 using Robust.Shared.Prototypes;
 
-namespace Content.Server._Exodus.Silicons.Borgs;
+namespace Content.Server._Exodus.Chemistry;
 
-public sealed class BorgParamedicHypoSystem : EntitySystem
+public sealed class CyclingHypospraySystem : EntitySystem
 {
     [Dependency] private readonly IPrototypeManager _prototype = default!;
     [Dependency] private readonly ReagentAutoRechargeSystem _reagentAutoRecharge = default!;
@@ -22,13 +21,13 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
 
     public override void Initialize()
     {
-        SubscribeLocalEvent<BorgParamedicHypoComponent, MapInitEvent>(OnMapInit);
-        SubscribeLocalEvent<BorgParamedicHypoComponent, UseInHandEvent>(OnUseInHand, before: [typeof(HypospraySystem)]);
-        SubscribeLocalEvent<BorgParamedicHypoComponent, ActivateInWorldEvent>(OnActivate);
-        SubscribeLocalEvent<BorgParamedicHypoComponent, AfterInteractEvent>(OnAfterInteract, before: [typeof(HypospraySystem)]);
+        SubscribeLocalEvent<CyclingHyposprayComponent, MapInitEvent>(OnMapInit);
+        SubscribeLocalEvent<CyclingHyposprayComponent, UseInHandEvent>(OnUseInHand, before: [typeof(HypospraySystem)]);
+        SubscribeLocalEvent<CyclingHyposprayComponent, ActivateInWorldEvent>(OnActivate);
+        SubscribeLocalEvent<CyclingHyposprayComponent, AfterInteractEvent>(OnAfterInteract, before: [typeof(HypospraySystem)]);
     }
 
-    private void OnMapInit(Entity<BorgParamedicHypoComponent> ent, ref MapInitEvent args)
+    private void OnMapInit(Entity<CyclingHyposprayComponent> ent, ref MapInitEvent args)
     {
         NormalizeSolution(ent);
 
@@ -39,7 +38,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
             _reagentAutoRecharge.SetReagent((ent, recharge), reagent);
     }
 
-    private void OnUseInHand(Entity<BorgParamedicHypoComponent> ent, ref UseInHandEvent args)
+    private void OnUseInHand(Entity<CyclingHyposprayComponent> ent, ref UseInHandEvent args)
     {
         if (args.Handled)
             return;
@@ -49,7 +48,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
         args.ApplyDelay = false;
     }
 
-    private void OnActivate(Entity<BorgParamedicHypoComponent> ent, ref ActivateInWorldEvent args)
+    private void OnActivate(Entity<CyclingHyposprayComponent> ent, ref ActivateInWorldEvent args)
     {
         if (args.Handled || !args.Complex)
             return;
@@ -58,7 +57,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
         args.Handled = true;
     }
 
-    private void OnAfterInteract(Entity<BorgParamedicHypoComponent> ent, ref AfterInteractEvent args)
+    private void OnAfterInteract(Entity<CyclingHyposprayComponent> ent, ref AfterInteractEvent args)
     {
         if (args.Handled || !args.CanReach || args.Target == null)
             return;
@@ -70,7 +69,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
         args.Handled = true;
     }
 
-    private void CycleReagent(Entity<BorgParamedicHypoComponent> ent, EntityUid user)
+    private void CycleReagent(Entity<CyclingHyposprayComponent> ent, EntityUid user)
     {
         if (ent.Comp.Reagents.Count == 0)
             return;
@@ -87,7 +86,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
             _reagentAutoRecharge.SetReagent((ent, recharge), reagent);
     }
 
-    private void NormalizeSolution(Entity<BorgParamedicHypoComponent> ent)
+    private void NormalizeSolution(Entity<CyclingHyposprayComponent> ent)
     {
         if (!_solutions.TryGetSolution(ent.Owner, ent.Comp.SolutionName, out var solutionEnt, out var solution))
             return;
@@ -95,7 +94,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
         NormalizeSolution(ent, solutionEnt.Value, solution);
     }
 
-    private void NormalizeSolution(Entity<BorgParamedicHypoComponent> ent, Entity<SolutionComponent> solutionEnt, Solution solution)
+    private void NormalizeSolution(Entity<CyclingHyposprayComponent> ent, Entity<SolutionComponent> solutionEnt, Solution solution)
     {
         if (!TryGetCurrentReagent(ent.Comp, out var reagent))
             return;
@@ -111,7 +110,7 @@ public sealed class BorgParamedicHypoSystem : EntitySystem
         _solutions.TryAddReagent(solutionEnt, reagent, volume, out _);
     }
 
-    private bool TryGetCurrentReagent(BorgParamedicHypoComponent component, out ProtoId<ReagentPrototype> reagent)
+    private bool TryGetCurrentReagent(CyclingHyposprayComponent component, out ProtoId<ReagentPrototype> reagent)
     {
         reagent = default;
         if (component.Reagents.Count == 0)
